I have moved a lot o fthe folders and files around to organize them better to get ready for Routing the pages.  I have implemented React Router. Although I do not have anything as of yet to route to, it ppears to be working correctly by rendering the homepage just fine.  I had to condense the homepage components into a single component for React Router to work properly.  I am committing current changes, and I will begin adding the new pages to further test the router to be sure it is working correctly.
