+++ src/index.js
@@ -1,11 +1,5 @@
 
 import ReactDOM from 'react-dom';
-// import * as ReactDOM from 'react-dom/client';
 import App from './App.js';
 
 ReactDOM.render(<App />, document.getElementById('root'));
-
-// const container = document.getElementById('root');
-
-// const root = ReactDOM.createRoot(container);
-// root.render(<App />);
